refactor MLP model and update imports for improved structure
