+++ hardware_accelerators/nn/__init__.py
@@ -1,2 +1,2 @@
-from .mlp import model_factory
-from .util import get_pytorch_device
+from .util import get_pytorch_device, model_factory, load_model
+from .mlp import MLP
