+++ hardware_accelerators/nn/util.py
@@ -1,4 +1,5 @@
 import torch
+from .mlp import MLP
 
 
 def get_pytorch_device() -> torch.device:
@@ -10,3 +11,20 @@ def get_pytorch_device() -> torch.device:
         device = torch.device("cpu")
 
     return device
+
+
+INPUT_DIM = 28 * 28  # MNIST images are 28x28
+HIDDEN_SIZE = 128
+OUTPUT_DIM = 10
+
+
+def model_factory():
+    return MLP(INPUT_DIM, HIDDEN_SIZE, OUTPUT_DIM)
+
+
+def load_model(model_path: str):
+    device = get_pytorch_device()
+    model = model_factory()
+    model.to(device)
+    model.load_state_dict(torch.load(model_path, map_location=device))
+    return model
